Updated Readme, removed deprecated Manufacturing Calculator
